Add datasubmit.php to add data to json file and submit the data from script.babel
PHP is used for this version but it will be replaced by NodeJS in future versions.
The script.babel will send a post request and the score will be saved in json file,
right now file based data saving is enabled on introduction of NodeJS for future
versions the data will be handled by a Database based mechanism.